Extract MemPubSub from PubSub
Before implementing the distributed PubSub implementation using ETCD, we
need to extract the temporarily implemented MemoryPubSub and leave only
the high-level abstraction logic in PubSub.
