+++ yorkie/config.sample.json
@@ -1,20 +1,20 @@
 {
-    "RPC": {
-        "Port": 11101,
-        "CertFile": "",
-        "KeyFile": ""
-    },
-    "Metrics": {
-        "Port": 11102
-    },
-    "Mongo": {
-        "ConnectionTimeoutSec": 5,
-        "ConnectionURI": "mongodb://localhost:27017",
-        "YorkieDatabase": "yorkie-meta",
-        "PingTimeoutSec": 5
-    },
-    "Backend": {
-        "SnapshotThreshold": 500,
-        "SnapshotInterval": 100
-    }
+  "RPC": {
+    "Port": 11101,
+    "CertFile": "",
+    "KeyFile": ""
+  },
+  "Metrics": {
+    "Port": 11102
+  },
+  "Mongo": {
+    "ConnectionTimeoutSec": 5,
+    "ConnectionURI": "mongodb://localhost:27017",
+    "YorkieDatabase": "yorkie-meta",
+    "PingTimeoutSec": 5
+  },
+  "Backend": {
+    "SnapshotThreshold": 500,
+    "SnapshotInterval": 100
+  }
 }
